+++ src/App.js
@@ -1,28 +1,13 @@
 import React from "react";
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
-const todoList = [
-  {
-    id: 1,
-    title: "react project",
-  },
-  {
-    id: 2,
-    title: "JavaScript XML",
-  },
-  {
-    id: 3,
-    title: "react dove class",
-  },
-];
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (list) {
-          return <li key={list.id}>{list.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
