+++ src/TodoList.js
@@ -1,0 +1,29 @@
+import React from "react";
+
+const todoList = [
+  {
+    id: 1,
+    title: "react project",
+  },
+  {
+    id: 2,
+    title: "JavaScript XML",
+  },
+  {
+    id: 3,
+    title: "react dove class",
+  },
+];
+
+function TodoList() {
+  return (
+    <div>
+      <ul>
+        {todoList.map(function (list) {
+          return <li key={list.id}>{list.title}</li>;
+        })}
+      </ul>
+    </div>
+  );
+}
+export default TodoList;
